# src/Airslip.Analytics.Api/openapi/analytics.api.2021.11.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Internal Analytics API",
    "description": "An internal only API used for any UI products",
    "version": "2021.11"
  },
  "servers": [
    {
      "url": "http://localhost:7101"
    }
  ],
  "paths": {
...
    "/2021.11/snapshot/{snapshotType}": {
      "post": {
        "tags": [
          "Snapshot"
        ],
        "operationId": "Get Snapshot",
        "parameters": [
          {
            "name": "snapshotType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DashboardSnapshotType"
            }
          },
          {
            "name": "dayRange",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 7
            }
          },
          {
            "name": "statRange",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "integrationId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedSnapshotSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSnapshotModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/2021.11/snapshot/revenue": {
      "post": {
        "tags": [
          "Snapshot"
        ],
        "operationId": "Get Revenue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedSeriesSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardGraphSeriesModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/2021.11/snapshot/cashflow": {
      "post": {
        "tags": [
          "Snapshot"
        ],
        "operationId": "Get Cashflow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedSeriesSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardGraphSeriesModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AirslipUserType": {
        "enum": [
          "Unknown",
          "Standard",
          "Merchant",
          "Unregistered",
          "Partner",
          "Administrator",
          "InternalApi"
        ],
        "type": "string"
      },
...
      "DashboardGraphSeriesModel": {
        "required": [
          "currencyCode",
          "endDate",
          "series",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "currencyCode": {
            "type": "string"
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Series"
            }
          }
        },
        "additionalProperties": false
      },
      "DashboardSnapshotModel": {
        "required": [
          "balance",
          "currencyCode",
          "dayRange",
          "metrics",
          "movement",
          "timeStamp"
        ],
        "type": "object",
        "properties": {
          "balance": {
            "type": "number",
            "description": "A description about the property should go here",
            "format": "double"
          },
          "timeStamp": {
            "type": "integer",
            "description": "A description about the property should go here",
            "format": "int64"
          },
          "movement": {
            "type": "number",
            "description": "A description about the property should go here",
            "format": "double"
          },
          "currencyCode": {
            "type": "string",
            "description": "A description about the property should go here"
          },
          "metrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnapshotMetric"
            },
            "description": "A description about the property should go here"
          },
          "dayRange": {
            "type": "integer",
            "description": "A description about the property should go here",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DashboardSnapshotType": {
        "enum": [
          "TotalSales",
          "TotalRefunds",
          "CurrentBalance"
        ],
        "type": "string"
      },
...
      "ErrorResponse": {
        "required": [
          "errorCode",
          "links",
          "metadata"
        ],
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": { }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        },
        "additionalProperties": false
      },
...
      "Link": {
        "required": [
          "href",
          "method",
          "rel"
        ],
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "rel": {
            "type": "string"
          },
          "method": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "OwnedSeriesSearchModel": {
        "required": [
          "currencyCode",
          "endDate",
          "ownerAirslipUserType",
          "ownerEntityId",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "integrationId": {
            "type": "string",
            "nullable": true
          },
          "ownerEntityId": {
            "type": "string"
          },
          "ownerAirslipUserType": {
            "$ref": "#/components/schemas/AirslipUserType"
          },
          "currencyCode": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "OwnedSnapshotSearchModel": {
        "required": [
          "currencyCode",
          "ownerAirslipUserType",
          "ownerEntityId"
        ],
        "type": "object",
        "properties": {
          "ownerEntityId": {
            "type": "string"
          },
          "ownerAirslipUserType": {
            "$ref": "#/components/schemas/AirslipUserType"
          },
          "currencyCode": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PeriodType": {
        "enum": [
          "Minute",
          "Hour",
          "Day",
          "Week",
          "Month",
          "Year"
        ],
        "type": "string"
      },
...
      "Series": {
        "required": [
          "data",
          "metrics",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "metrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelyMetric"
            }
          },
          "data": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "additionalProperties": false
      },
      "SnapshotMetric": {
        "required": [
          "balance",
          "timeStamp"
        ],
        "type": "object",
        "properties": {
          "timeStamp": {
            "type": "integer",
            "format": "int64"
          },
          "balance": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "TimelyMetric": {
        "required": [
          "balance",
          "description",
          "period",
          "periodType"
        ],
        "type": "object",
        "properties": {
          "period": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "periodType": {
            "$ref": "#/components/schemas/PeriodType"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "e.g Bearer Api_Key.\n\nRequests to the Airslip API are authenticated using the applications `Api_Key`. You can view and manage your credentials in the Airslip Dashboard.\n\nAn Api_Key provides connectivity to all authenticated Airslip API endpoints, so it is important to keep these credentials secure. Do not share your Api_Key in publicly accessible areas such as GitHub, client-side code, or easily accessible configuration settings.\n\nAuthentication is performed using Bearer Authentication. Your Api_Key should be sent as the token.\n\nAll requests should be made via HTTPS.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}